+++ Editor/Common/Controls/ScalarBox.cs
@@ -1,0 +1,12 @@
+﻿using System.Windows;
+
+namespace Editor.Common.Controls
+{
+    public class ScalarBox : NumberBox
+    {
+        static ScalarBox()
+        {
+            DefaultStyleKeyProperty.OverrideMetadata(typeof(ScalarBox), new FrameworkPropertyMetadata(typeof(ScalarBox)));
+        }
+    }
+}