+++ Editor/Common/Enums/VectorType.cs
@@ -1,0 +1,9 @@
+﻿namespace Editor.Common.Enums
+{
+    public enum VectorType
+    {
+        Vector2,
+        Vector3,
+        Vector4
+    }
+}